extracted Runtime method to a new class
