+++ mRemoteV1/mRemoteV1.csproj
@@ -405,6 +405,7 @@
     <Compile Include="UI\Controls\ToolStripSplitButton.cs">
       <SubType>Component</SubType>
     </Compile>
+    <Compile Include="UI\FontOverrider.cs" />
     <Compile Include="UI\FormExtensions.cs" />
     <Compile Include="UI\Forms\CompositeCredentialRepoUnlockerForm.cs">
       <SubType>Form</SubType>
